+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.0
 =======
 
-* Removed configuration profile from the token server.
 * Added an example app to demonstrate Media Device Selection.
+* Updated twilio-video to 1.0.0
+* Removed configuration profile
 * Consolidated the various twilio-video.js quickstarts into a single project.
-